mcp/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "a4s-mcp"
version = "0.1.0"
description = "MCP server for A4S"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk>=0.3.0",
    "httpx>=0.28",
    "mcp>=1.26.0",
    "pydantic-settings>=2.12.0",
]

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["a4s_mcp"]
